refactor: 상품 이벤트 매핑 로직을 Mapper로 분리
