+++ catalog-service/src/main/java/com/node5/catalogservice/product/application/port/ProductIndexEventPort.java
@@ -1,6 +1,6 @@
 package com.node5.catalogservice.product.application.port;
 
-import com.node5.catalogservice.product.event.ProductIndexEvent;
+import com.node5.common.event.ProductIndexEvent;
 
 public interface ProductIndexEventPort {
 	void publish(ProductIndexEvent event);
